comment.py 댓글 번호, 날짜 추가, 댓글 삭제 api 추가

+++ comment.py
@@ -11,7 +11,8 @@ def comment_post():
     name_receive = request.form["name_give"]
     comment_receive = request.form["comment_give"]
     num_receive = request.form['m_id_give']
-    password_receive = request.form['password']
+    password_receive = request.form['password_give']
+    date_receive = request.form['date_give']
     comments_list = list(db.comments.find({}, {'_id': False}).sort('c_num'))
     # count = len(comments_list) + 1
 
@@ -24,24 +25,31 @@ def comment_post():
         'c_id': name_receive,
         'c_num': last_comments_num + 1,
         'comment': comment_receive,
-        'password': password_receive
+        'password': password_receive,
+        'date': date_receive
     }
 
     db.comments.insert_one(doc)
     return jsonify({'msg': '댓글 완료!'})
 
 
-@app.route("/comments/", methods=["DELETE"])
-def comments_del():
+
+def comment_del():
     c_num_receive = request.form['c_num_give']
     password_receive = request.form['password']
     # db.comments_list
 
     # user = db.users.find_one({'id': id_receive}, {'_id': False})
 
-    comment = db.comments.find_one({'c_num': c_num_receive}, {'_id': False})
+    comment = db.comments.find_one({'c_num': int(c_num_receive)}, {'_id': False})
 
-    db.comments.delete_one({'c_num': int(c_num_receive)})
+    if comment['password'] == password_receive :
+        db.comments.delete_one({'c_num': int(c_num_receive)})
+        return jsonify({'delete':'success', 'msg': '삭제 성공'})
+    else :
+        return jsonify({'delete':'fail', 'msg': '비밀번호가 틀립니다!'})
+
+
     #
     # comments_list = list(db.comments.find({}, {'_id': False}))
 
@@ -51,6 +59,6 @@ def comments_del():
     #     rows = i['num']
     #     db.bucket.update_one({'num': rows}, {'$set': {'num': index}});
 
-    return jsonify({'msg': '버킷 삭제!'})
+
